pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = "create an app that has just one single api - which returns \"Hello World\""
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pyjwt = "*"
asyncio = "*"
bcrypt = "*"
fastapi = "*"
h11 = "*"
prisma = "*"
pydantic = "*"
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
